+++ packages/reacord/library.new/core/button-shared-props.ts
